officially support negative increment on grk_buf offset

+++ src/lib/jp2/util/util.h
@@ -83,7 +83,7 @@ struct grk_buf {
 		len(length),
 		owns_data(ownsData) {}
 	~grk_buf();
-	void incr_offset(uint64_t off);
+	void incr_offset(ptrdiff_t off);
 	uint8_t* curr_ptr();
 
 	uint8_t *buf;		/* internal array*/
